move isCol1 to common/device
